refactor: decompose app.js into es modules
