+++ index.html
@@ -90,7 +90,7 @@
     </div>
 
     <script src="vendor/papaparse.min.js"></script>
-    <script src="app.js"></script>
+    <script type="module" src="app.js"></script>
     <script data-goatcounter="https://butteredgroove.goatcounter.com/count" async src="//gc.zgo.at/count.js"></script>
   </body>
 </html>
