+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fac-flipper",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "private": true,
   "devDependencies": {
     "@biomejs/biome": "1.8.3",
